Move bgs_itf_service to itf controller
